# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Greater Bayelsa leadership development platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dues/status:
    get:
      operationId: getDuesStatus
      tags: [dues]
      summary: Per-member dues status for a period within scope
      parameters:
        - name: period
          in: query
          required: true
          schema: { type: string }
        - name: villageId
          in: query
          required: false
          schema: { type: integer }
        - name: unitId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Member dues statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberDuesStatus"
  /dues/rollup:
    get:
      operationId: getDuesRollup
      tags: [dues]
      summary: Aggregated paid/overdue rollup by village and unit within scope
      parameters:
        - name: period
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Rollup rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DuesRollupRow"
  /dues/bulk:
    post:
      operationId: recordBulkDues
      tags: [dues]
      summary: Record a bulk cash payment for unit members (unit leader)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDuesBody"
      responses:
        "201":
          description: Payments recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDuesResult"
  /dues/reminders:
    post:
      operationId: sendDuesReminders
      tags: [dues]
      summary: Send in-platform dues reminders to overdue members (HQ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuesRemindersBody"
      responses:
        "200":
          description: Reminders sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuesRemindersResult"

  # ---------- Meetings ----------
...
components:
  schemas:
...
    MemberDuesStatus:
      type: object
      required: [userId, firstName, lastName, membershipCode, paid]
      properties:
        userId: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        membershipCode: { type: string }
        unitId: { type: [integer, "null"] }
        villageId: { type: [integer, "null"] }
        paid: { type: boolean }
        method: { type: [string, "null"] }
    DuesRollupRow:
      type: object
      required: [villageId, villageName, paidCount, totalCount]
      properties:
        villageId: { type: integer }
        villageName: { type: string }
        unitId: { type: [integer, "null"] }
        unitName: { type: [string, "null"] }
        paidCount: { type: integer }
        totalCount: { type: integer }
    BulkDuesBody:
      type: object
      required: [userIds, period, reference]
      properties:
        userIds:
          type: array
          minItems: 1
          items: { type: integer }
        period: { type: string }
        reference: { type: string, minLength: 3 }
        receiptUrl: { type: [string, "null"] }
    BulkDuesResult:
      type: object
      required: [recorded, alreadyPaid]
      properties:
        recorded: { type: integer }
        alreadyPaid: { type: integer }
    DuesRemindersBody:
      type: object
      required: [period]
      properties:
        period: { type: string }
    DuesRemindersResult:
      type: object
      required: [notified]
      properties:
        notified: { type: integer }
...